requirements.txt
llama-cpp-python==0.2.90
streamlit==1.40.0
faiss-cpu==1.9.0.post1
pypdf==5.1.0
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.67.1
requests==2.32.3
python-dateutil==2.9.0

# PyTorch and ML libraries - TESTED COMPATIBLE VERSIONS
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.38.2
sentence-transformers==2.5.1
